requirements.txt
pycryptodome
psutil
pyqt5